+++ environments/docker/install.sh
@@ -1,20 +1,24 @@
 #!/bin/bash
 set -euo pipefail
 
-# rasterio doesn't have linux/arm64 wheels, so if this platform target is
-# linux/arm64 then don't install this optional dependency (attempting to
-# build it rather than install the wheel also fails).
+# rasterio doesn't provide linux/arm64 wheels, so if the target platform is
+# linux/arm64, don't install this optional dependency (attempting to build
+# it rather than install the wheel will also fail).
+# see https://github.com/rasterio/rasterio-wheels/issues/69
 if [[ "$TARGETPLATFORM" == "linux/arm64" ]]
 then
-  NOEXTRA="--no-extra=raster"
+  NOEXTRA="--no-extra=raster --no-extra=all"
 else
   NOEXTRA=""
 fi
 
+# install uv and add the executable to the PATH
 wget -qO- https://astral.sh/uv/0.7.12/install.sh | sh
 source $HOME/.local/bin/env
-uv export --no-cache --no-build --no-dev --all-extras $NOEXTRA > reqs.txt
-uv pip install --no-cache --no-build --system --strict -r reqs.txt -r requirements-extras.txt
+
+# install all requirements into the existing system environment
+uv export --no-cache --no-build --no-dev --all-extras $NOEXTRA > requirements.txt
+uv pip install --no-cache --no-build --system --strict -r requirements.txt
 uv cache clean
 uv pip check
 uv pip list